+++ src/components/QuestionEdit.js
@@ -42,10 +42,16 @@ import Card from '@material-ui/core/Card'
 import CardActions from '@material-ui/core/CardActions'
 import CardContent from '@material-ui/core/CardContent'
 import Divider from '@material-ui/core/Divider'
+import IconButton from '@material-ui/core/IconButton'
 import InputAdornment from '@material-ui/core/InputAdornment'
 import Radio from '@material-ui/core/Radio'
 import { makeStyles } from '@material-ui/core/styles'
 import TextField from '@material-ui/core/TextField'
+
+//----------------------------------------------------------------//
+// Material UI Icons
+//----------------------------------------------------------------//
+import DeleteIcon from '@material-ui/icons/Delete'
 
 //----------------------------------------------------------------//
 // Custom Class Styles
@@ -121,7 +127,18 @@ export default ({ answerRefs, optionRefs, question = {question:'', options:['','
                       inputRef={() => answerRefs.current[questionIndex] = selectedValue}
                     />
                   </InputAdornment>
+                ),
+                endAdornment: (
+                  <InputAdornment position='end'>
+                    <IconButton
+                      aria-label='delete option'
+                      edge='end'
+                    >
+                      <DeleteIcon />
+                    </IconButton>
+                  </InputAdornment>
                 )
+                
               }}
               inputRef={value => optionRefs.current[questionIndex][optionIndex] = value}
               label={`Option ${String.fromCharCode(65 + optionIndex)}`}